intergrate with main Activity
