+++ Android/app/src/main/AndroidManifest.xml
@@ -64,18 +64,12 @@
             android:name=".View.StartActivity"
             android:label="@string/title_activity_start"
             android:theme="@android:style/Theme.NoTitleBar" >
-
-        </activity>
-        <activity
-            android:name=".View.QRActivity"
-            android:label="@string/title_activity_qr"
-
-            android:theme="@style/AppTheme.NoActionBar" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+
         </activity>
 
         <activity android:name="com.google.zxing.client.android.CaptureActivity"
