+++ Android/app/src/main/java/org/secmem/swm_online/View/MainActivity.java
@@ -15,6 +15,7 @@ import android.webkit.WebViewClient;
 import android.widget.Button;
 import android.widget.ProgressBar;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GooglePlayServicesUtil;
@@ -29,7 +30,9 @@ public class MainActivity extends Activity {
     private static final int PLAY_SERVICES_RESOLUTION_REQUEST = 9000;
     private static final String TAG = "MainActivity";
     private static final String URL = "http://www.swmem.org";
+    private static int REQUEST_RENT = 0;
 
+    private Button qrButton;
     private Button mRegistrationButton;
     private ProgressBar mRegistrationProgressBar;
     private BroadcastReceiver mRegistrationBroadcastReceiver;
@@ -47,8 +50,9 @@ public class MainActivity extends Activity {
 
         mRegistrationProgressBar = (ProgressBar) findViewById(R.id.registrationProgressBar);
         mRegistrationProgressBar.setVisibility(ProgressBar.GONE);
-
+        qrButton =  (Button) findViewById(R.id.qr_btn);
         mRegistrationButton = (Button) findViewById(R.id.registrationButton);
+
         mRegistrationButton.setOnClickListener(new View.OnClickListener() {
             /**
              * @param view
@@ -59,8 +63,24 @@ public class MainActivity extends Activity {
             }
         });
 
-        getInstanceIdToken();
+        qrButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent("com.google.zxing.client.android.SCAN");
 
+                intent.putExtra("SCAN_MODE", "QR_CODE_MODE");
+
+                startActivityForResult(intent, REQUEST_RENT);
+
+
+            }
+        });
+
+
+        //걍 실행되자마자 토큰 갖어오게 해둠
+       // getInstanceIdToken();
+
+        // 웹뷰 설정
         myWebView = (WebView) findViewById(R.id.webview);
         myWebView.setWebViewClient(new WebViewClient());
         WebSettings webSettings = myWebView.getSettings();
@@ -169,6 +189,35 @@ public class MainActivity extends Activity {
 //        }
 //        return super.onKeyUp(keyCode, event);
 //    }
+
+
+
+    /*
+        QRcode 핸들러 정의
+
+     */
+
+
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+
+        if(requestCode == REQUEST_RENT) {
+
+            if(resultCode == Activity.RESULT_OK)
+            {
+                String contents = data.getStringExtra("data");
+                //위의 contents 값에 scan result가 들어온다.
+
+                Toast.makeText(getApplicationContext(), contents, Toast.LENGTH_SHORT).show();
+                mInformationTextView.setText(contents);
+            }
+
+        }
+
+        super.onActivityResult(requestCode, resultCode, data);
+    }
+
+
+
 
     public class MyWebViewClient extends WebViewClient {
